+++ internal/handler/list_notifications.go
@@ -6,13 +6,13 @@ import (
 	"net/http"
 	"strconv"
 
-	"github.com/Kilat-Pet-Delivery/lib-common/auth"
-	"github.com/Kilat-Pet-Delivery/lib-common/middleware"
-	"github.com/Kilat-Pet-Delivery/lib-common/response"
-	"github.com/Kilat-Pet-Delivery/lib-proto/dto"
-	"github.com/Kilat-Pet-Delivery/service-notification/internal/repository"
 	"github.com/gin-gonic/gin"
 	"github.com/google/uuid"
+	"github.com/niaga-labs/niaga-labs-pet-lib-common/auth"
+	"github.com/niaga-labs/niaga-labs-pet-lib-common/middleware"
+	"github.com/niaga-labs/niaga-labs-pet-lib-common/response"
+	"github.com/niaga-labs/niaga-labs-pet-lib-proto/dto"
+	"github.com/niaga-labs/niaga-labs-pet-service-notification/internal/repository"
 	"go.uber.org/zap"
 )