+++ internal/repository/notification_list_repo.go
@@ -8,8 +8,8 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/Kilat-Pet-Delivery/lib-proto/dto"
 	"github.com/google/uuid"
+	"github.com/niaga-labs/niaga-labs-pet-lib-proto/dto"
 	"gorm.io/gorm"
 )